Assistant checkpoint: Update portion count input to handle decimal numbers
Assistant generated file changes:
- client/src/components/menu/PreparationSheet.tsx: Update portion count to handle decimal numbers

---

User prompt:

Also, I can't seem to type float numbers in the portion field

+++ client/src/components/menu/PreparationSheet.tsx
@@ -454,7 +454,7 @@ export default function PreparationSheet({
                         step="0.01"
                         {...field}
                         onChange={(e) =>
-                          field.onChange(parseInt(e.target.value))
+                          field.onChange(parseFloat(e.target.value))
                         }
                       />
                     </FormControl>
